requirements.txt
jinja2>=3.1.2
pyyaml>=6.0
pypdf>=4.0.0
